Fix: app state sometimes not saved on exit
The method JuK::saveConfig() was sometimes not called when the app
exited, causing user-initiated changes to be lost e.g. whether or not
the system tray should be used.

Also, force the "StartDocked" config parameter to false, making the main
window always visible on application start. Starting the app in minimized
state is confusing if you don't immediately notice the system tray icon.