App: selectively rename 'juk' to 'jukebox'
Change the executable name to 'jukebox', and change the app name used
in the main window title and in the Help Menu. Internally, we are still
'juk' as far as the KDE4 paths, icons and config files are concerned.

Also update our KDE4 launcher file, jukebox.desktop, with the new info.
The goal is to allow this app to be installed at the same time as original
juk app, without causing conflict.