Playlist: Fix the checkmark state on the View|Show Columns menu.
The checkmarks on the "View|Show Columns" menu did not correctly reflect
the Playlist table columns which were shown or hidden. One problem is that
a separate KMenu was created for each Playlist that exists, and you
could never be sure which KMenu the "Show Columns" Action pointed to. Fix
this by creating only a single KMenu, and using it everywhere.