File Menu: improve item text, and add StatusBar tips
The commands on the File are too general, and it's hard to determine
what a command will actually do. Append the text " Playlist" to explicitly
specify the commands which operate on playlists, and add status bar text to
provide additional information about what will happen.

The "Open..." item text was changed to "Import Playlist..." to more
accurately specify what it does, and the Control-O accelerator was removed.