Add a dirty bit to the Playlist object to remember if any files in the .m3u list have been added/deleted/renamed since it was created, read from disk or last saved.
Add a new Playlist::saveFile() method to write .m3u file to disk, which
also returns status.

Add a new PlaylistBox::savePlaylistsToDisk() to save all modified playlists;
it takes an argument to determine whether the user should be prompted for
Yes/No to save each dirty playlist.  If the desktop is logging out, the dialog
boxes should be disabled.

Prior to this patch, the app silently discarded any playlist changes when the
app exited. The new default policy is "save the changes".