playlist: replace Playlist::readOnly() by more fine-grained permissions
readOnly() is very, very restrictive. This property affects what THE USER
is allowed to do from the menus and keyboard. It:
- prohibits any changes to the playlist content
- prohibits deletion of the playlist itself
- prohibits renaming the playlist
- prohibits saving the playlist to disk

It's so restrictive that it's almost useless. Replace it by three new
finer-grain methods, which we'll take full advantage of in a future patch:

Playlist::canDelete()         # delete the playlist and the .m3u file itself
Playlist::canModifyContent()  # add,delete,rename within the list
Playlist::canRename()         # edit the label on playlist icon

This patch also replaces all calls to readOnly() by the proper new
method. The visible application behavior should be changed.