menu: change "Save Playlist As..." to "Export Playlist..."
By removing the SaveAs menu item, we avoid changing the association
between a specific playlist icon and it's underlying .m3u file.

For consistency, rename method slotSaveAs() to slotExportFile()
and saveAs() to exportFile().

The action label "file_save_as" remains unchanged to maximize compatibility
with jukui.rc, which we still rely on.