fix: the app loses track of .m3u files external to 'managed folder' tree(s)
In theory, the app caches the playlists that are shown in the left panel.
But a serious logic error in Cache::loadPlaylists() caused the cached data
for most playlists to be discarded.

At app startup, checking if the playlist filename is present in the
PlaylistCollection hash is wrong for two reasons:
- the m3u fileName was just added to the hash by the operator>> immediately
  preceding the check. So if there's a fileName, the check returns true.
- the folder scanner has not run yet, so the m3u filename can't possibly
  be in the hash list for that reason.

So what the bad code *actually* does is to discard any cached playlist
information that also stored an m3u filename. This is not the behavior
we need.