import: improve import dialog filtering options
The filter options for the import dialog did not allow *.m3u files to
be selected, and did not give an 'All Files (*.*)' choice either.

This patch adds both choices to the filter pulldown, and makes *.m3u the
default.

The trendy thing seems to be to use mime types for dialog filters, but in
KDE 4.10.5, the description for *.m3u file type is "MP3 audio (streamed)".
That's crap.

So instead, go 'old school': use setFilter() instead of setMimeFilter(),
and provide a text description that makes sense. If non-audio files are
chosen by the user, that'll be detected during import processing and
the bad files will be ignored.