menu: eliminate the special menu "Hide" logic for 'Play Queue'
The Play Queue Playlist, internally named UpcomingPlaylist, got special
treatment on the File Menu. The 'Remove Playlist...' item changed to 'Hide',
trying to convey the idea that the playlist would not actually be deleted.

Instead of doing something special, just rely on the existing 'View|Show
Play Queue' Menu item, where clearing the check mark already implies that
it'll be hidden by the user action. Leave the 'Remove Playlist' item on
the menu alone, eliminating the need to explain an exception.

+1 for simplifying the GUI.