bug fix: at app startup, volume level not set correctly
As discussed in bugs.kde.org #321172, Phonon ignores setVolume() when not
actually in the playing state. This caused our attempt to set the
volume level at startup to not be successful. Our GUI Volume widget showed
the desired value, but internally Phonon was still using 100%.

In Phonon v4.7 this was fixed. To give us compatibility with
earlier versions of Phonon, have our PlayerManager save the desired volume
level and apply it after Phonon enters the playing state.