volume: make our in-app volume mute behavior match what kmix does
The Phonon backend has a dedicated feature to mute/unmute the audio. But for
whatever reason, kmix doesn't use it. Instead, clicking the mute button
lowers the volume to zero, and unmute restores it to the previous value.

It's odd that to see one mute behavior from the VolumePopupButton, and a
different mute behavior from the slider in kmix in the system tray.

Change the VolumePopupButton widget to use the lower-volume-to-zero
approach.