playlist: add isContentMutable() flag
This flag determines if the playlist content can be modified or not by
the app. This is _state_ that can be changed at any time. This flag is
distinct from canModifyContent(), which applies only to USER ACTION.

This will be used for playlists read from read-only .m3u files. Also
implement setContentMutable(bool), which is currently unused.