fix: backtrace at end of final track of a playlist
The message is 'Trying to add empty file, backtrace follows:...'

It is from the Tag constructor. StatusLabel::updateCurrent() is called at the
end of the final track in a playlist. If playlist looping is disabled, at that
point in time PlaylistInterface::currentFile() returns FileHandle::null, so the
attempt to invoke FileHandle::tag() goes badly.

The fix is to check for FileHandle::null, and bail out if it happens.