volume: yet another attempt to fix a too-loud volume at startup
The code to initialize volume by calling PlayerManager::setVolume() seems to work
reliably. However, the change I made in:

commit 7a7c86931d3d115b530e8ad4ce45dc56adbf6e71
"volume: fix for volume level not set correctly"

turns out not to be reliable. It seemed to work during my tests, but on later
days volume for first track played was 100% volume rather than the value
passed in to PlayerManager::setVolume().

This patch tries a different tack: set the volume on entry to BufferingState and again
on entry to PlayingState. It works so far. We'll see if it lasts.