fix: CollectionList was incorrectly allowing duplicate tracks
Yesterday I unintentionally commited an experiment where the Playlist
constructor defaulted to 'allow duplicate track' entries. In the big picture,
this change seems like a good idea, as the only Playlist that really seems to
need the restriction is CollectionList.

This patch restores the 'no dups' behavior for CollectionList.

Here is the commit where I broke it:

commit 4c36fdfde6d03e1307b369b4ffe1c6f74dbb858a
Author: Mike Scheutzow <mjs973@gmail.com>
Date:   Fri Sep 25 20:19:46 2015 -0400

    playlist: change default sort to 'insert order'