refactor: PlayerManager::stop() and StoppedState handling
Adding a m_bStopRequested flag tells us precisely what to do if
the current MediaObject enters the StoppedState.

Also gather the GUI update code into a playerHasStopped() method.
This code really ought to move out of this class, and be done by
a signalStop() listener. These updates now happen after the
MediaObject actually enters the StoppedState, rather than when
the stop() call is issued.

One advantage is that we don't have to wait 2 seconds to see if
something is going to happen; we can emit signalStop() immediately.
It much nicer to see the widgets update quicker.