player: combine signalPlay() and signalItemChanged()
Replace signalPlay() by signalItemChanged() in the few places it
was used. As with signalPlay(), issue signalItemChanged() when player
enters the PlayingState, and do not issue it at the end of playlist;
listeners can monitor signalStop() if they need that information.

This feels much cleaner, and avoids issuing a flurry of cpu-hogging
signals just when the audio decoder is trying to quickly fill the
audio buffer.