fix: UpcomingSequenceIterator deletes wrong UpcomingPlaylist Item
The user is allowed to use drag/drop to reorder the tracks in the
UpcomingPlaylist. So assuming that the Row 0 Item is the track which just
finished playing is wrong.

If the current() PlaylistItem is not in row 0, just leave the Item where
ever the user dragged it, and choose Row 0 as the next track.