fix: StatusLabel not updating consistently, part 2
In commit a0893dbe "fix: StatusLabel not updating consistently" I
tried to fix this. What I failed to notice at that time is that this
StatusLabel "freeze-up" happened the first time a track got dropped into
the Playlist Queue. In somebody's twisted world, this non-intuitive
behavior is a feature. Not in mine, so it's gone.