refactor: comment out useless URL Drop code
This code makes no sense. It is invoked when one or more URLs are dropped
onto the track table. But for that to happen, the track table must
already be visible. So the playlist icon is already selected. So this
code accomplishes nothing.

Commented rather than deleted in case we find a situation which
actually needs it.