fix: TrackSequenceManager::setCurrentPlaylist() has a split personality
setCurrentPlaylist() is being used for two different purposes, and the
resulting behavior makes no sense at all.

Feature 1 is to remember the playlist
associated with m_playNextItem, and connect some signals to it to
monitor for deletion before we get to it.

Feature 2 is to provide a default for the iterator in case
it can't decide which track to play.

These two things are independent of each other. This patch separates
the code to properly implement these two features. It also properly
disconnects the signals when they are no longer needed.