feat: implement getType() and getPolicy() for TreeViewItemPlaylist
TreeViewItemPlaylist are used to implement the leaf nodes under
CollectionList in the PlaylistBox. The benefit of this change is the
user can not edit the search pattern, and can not delete the playlist.

Prior to this patch, these two methods were inherited from
SearchPlaylist, which allows both of these actions.