feat: map PlaylistItem::ColumnType directly to table
The new assumption is that extra columns are added at the right end
of the table. This means that PlaylistItem::ColumnType directly maps
as Q3ListView column number, which gives much cleaner code. This change
led to changes in only Playlist::calculateColumnWeights() and the
PlaylistHistory class.

Another new assumption is that derived classes are responsible for
calling addColumn() themselves.