playlist: compare disk file content to cache to set modified flag
When we have both a particular m3u file in both disk and cache,
compare the individual file names to determine if the cached playlist
is modified or not. Prior to this patch, this app assumed the file
with the later timestamp was more recent. (Juk-2015-02-21 assumes
the disk-file is more recent.)