feat: don't save all playlists to cache
If we can get all the playlist info from the .m3u file, then don't
write the playlist to the cache. This is also a win at app startup,
because we won't spend time comparing the two lists.

For backward compatibility with Juk, the cache-reading code can't
assume this optimization has been done.